+++ common/src/main/java/com/lowdragmc/photon/client/gameobject/emitter/trail/TrailConfig.java
@@ -139,6 +139,7 @@ public class TrailConfig implements IPersistedSerializable {
         public void prepareStatus() {
             if (renderer.isBloomEffect()) {
                 beginBloom();
+                BloomEffect.setBloomColor(renderer.getBloomColor());
             } else {
                 RenderSystem.setShader(GameRenderer::getParticleShader);
             }