+++ common/src/main/java/com/lowdragmc/photon/client/gameobject/particle/TileParticle.java
@@ -587,16 +587,6 @@ public class TileParticle implements IParticle {
     }
 
     public void renderInternal(@Nonnull VertexConsumer buffer, Camera camera, float partialTicks) {
-        // set bloom color
-        if (config.renderer.isBloomEffect()) {
-            int bloomColor = config.renderer.getBloomColor().get(emitter.getRandomSource(), emitter.getT(partialTicks)).intValue();
-            float r = ColorUtils.red(bloomColor);
-            float g = ColorUtils.green(bloomColor);
-            float b = ColorUtils.blue(bloomColor);
-            float a = ColorUtils.alpha(bloomColor);
-            BloomEffect.setBloomColor(new Vector4f(r, g, b, a));
-        }
-
         var vec3 = camera.getPosition();
 
         var localPos = getLocalPos(partialTicks).mulPosition(getSpaceTransform());