+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v1.1.14
 * Fix incorrect shader attributes layout (thanks to @RogoShum)
+* Fix the bloom color crash with parallel rendering